+++ pakyow-core/lib/pakyow/actions/normalizer.rb
@@ -1,5 +1,7 @@
 # frozen_string_literal: true
 
+require "uri"
+
 require "pakyow/support/core_refinements/string/normalization"
 
 module Pakyow
@@ -9,6 +11,12 @@ module Pakyow
     class Normalizer
       using Support::Refinements::String::Normalization
 
+      def initialize
+        if canonical_uri = Pakyow.config.normalizer.canonical_uri
+          configure_canonical_uri!(canonical_uri)
+        end
+      end
+
       def call(connection)
         if strict_https? && require_https? && !https?(connection)
           unless http_allowed?(connection)
@@ -16,6 +24,8 @@ module Pakyow
           end
         elsif strict_https? && !require_https? && https?(connection)
           redirect!(connection, "http://#{File.join(connection.authority, connection.fullpath)}")
+        elsif strict_host? && !canonical?(connection)
+          redirect!(connection, "#{connection.scheme}://#{File.join(@canonical_uri.host, connection.fullpath)}")
         elsif strict_www? && require_www? && !www?(connection) && !subdomain?(connection)
           redirect!(connection, File.join(add_www(connection), connection.fullpath))
         elsif strict_www? && !require_www? && www?(connection)
@@ -26,6 +36,17 @@ module Pakyow
       end
 
       private
+
+      def configure_canonical_uri!(canonical_uri)
+        @canonical_uri = if canonical_uri.is_a?(URI)
+          canonical_uri
+        else
+          URI(canonical_uri)
+        end
+
+        Pakyow.config.normalizer.require_https = @canonical_uri.scheme == "https"
+        Pakyow.config.normalizer.require_www = false
+      end
 
       def redirect!(connection, location)
         connection.status = 301
@@ -89,6 +110,14 @@ module Pakyow
 
       def http_allowed?(connection)
         Pakyow.config.normalizer.allowed_http_hosts.include?(connection.host)
+      end
+
+      def strict_host?
+        instance_variable_defined?(:@canonical_uri)
+      end
+
+      def canonical?(connection)
+        connection.host == @canonical_uri.host
       end
     end
   end