+++ pakyow-core/CHANGELOG.md
@@ -21,6 +21,7 @@
     * `normalizer.allowed_http_hosts`: Array of hosts that are allowed as http (e.g. localhost)
   * [added] Require https by default when running in production
   * [fixed] Query string missing from normalized uris
+  * [added] Configure normalization through a canonical uri
 
 # 1.0.1